app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==3.2.2 
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
jinja2==3.1.4
aiosqlite==0.20.0